refactor: Extract task deadline logic and UI from EventTab into a new DeadlineSection component.
